requirements.txt
streamlit
scikit-learn
pandas
numpy
nltk
PyPDF2
python-docx